requirements-dev.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2
django-allauth==0.39.1
django-cors-headers==3.0.0
django-filter==2.1.0
django-rest-auth==0.9.5
djangorestframework==3.9.3
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
Markdown==3.1
mccabe==0.6.1
mysql-connector-python==8.0.13
mysqlclient==1.3.14
oauthlib==3.0.1
olefile==0.46
Pillow==6.0.0
protobuf==3.6.0
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
typed-ast==1.3.1
urllib3==1.24.3
wincertstore==0.2
wrapt==1.11.1
